+++ integration-libs/opf/base/root/model/index.ts
@@ -5,5 +5,6 @@
  */
 
 import './augmented-core.model';
-export * from './opf-configuration.model';
-export * from './opf.model';
+export * from './opf-base.model';
+export * from './opf-error-dialog.model';
+export * from './opf-metadata-store.model';
