+++ integration-libs/opf/base/root/model/opf-metadata-store.model.ts
@@ -1,0 +1,13 @@
+/*
+ * SPDX-FileCopyrightText: 2024 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export interface OpfMetadataModel {
+  termsAndConditionsChecked: boolean;
+  selectedPaymentOptionId: number | undefined;
+  defaultSelectedPaymentOptionId?: number;
+  isPaymentInProgress: boolean;
+  paymentSessionId: string | undefined;
+}
