+++ integration-libs/opf/base/root/services/opf-metadata-state-persistence.service.ts
@@ -8,7 +8,7 @@ import { Injectable, OnDestroy } from '@angular/core';
 import { StatePersistenceService } from '@spartacus/core';
 import { Observable, Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
-import { OpfMetadataModel } from '../model/opf.model';
+import { OpfMetadataModel } from '../model/opf-metadata-store.model';
 import { OpfMetadataStoreService } from './opf-metadata-store.service';
 
 /**
